Move file migration to a separate module
