+++ backend/src/cli/mod.rs
@@ -1,0 +1,3 @@
+mod upload_files;
+
+pub use self::upload_files::*;
