+++ backend/src/services/file_storage/s3_storage.rs
@@ -98,8 +98,4 @@ impl FileStorageInterface for S3FileStorage {
 
         Ok(())
     }
-
-    async fn find_files(&self) -> Result<Vec<u64>> {
-        Ok(Vec::new())
-    }
 }
