+++ backend/src/services/file_storage/trait.rs
@@ -6,5 +6,4 @@ use crate::types::Result;
 pub trait FileStorageInterface {
     async fn write_file(&self, id: u64, data: &[u8]) -> Result<()>;
     async fn read_file(&self, id: u64) -> Result<Vec<u8>>;
-    async fn find_files(&self) -> Result<Vec<u64>>;
 }
